% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_cmd.R
\name{bf_cmd}
\alias{bf_cmd}
\title{Get Path of BioFormat Command}
\usage{
bf_cmd(cmd)
}
\arguments{
\item{cmd}{Bioformat command}
}
\value{
A character string
}
\description{
Get Path of BioFormat Command
}
\examples{
bf_cmd("bfconvert")
}
