% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_parse_show_info.R
\name{bf_parse_show_info}
\alias{bf_parse_show_info}
\title{Parse output from \code{\link{showinf}}}
\usage{
bf_parse_show_info(file)
}
\arguments{
\item{file}{Output from \code{\link{showinf}},
a filename}
}
\value{
A list of output
}
\description{
Parse output from \code{\link{showinf}}
}
\examples{
file = "~/Downloads/2017_08_03__0006.czi"
if (file.exists(file)) {
res = showinf(file = file)
parsed = bf_parse_show_info(res)
res = showinf(file = file, debug = TRUE)
parsed2 = bf_parse_show_info(res)
testthat::expect_equal(parsed2$core, parsed$core)
testthat::expect_equal(parsed2$`series #0`, parsed$`series #0`)
}
}
