% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfconvert.R
\name{bfconvert}
\alias{bfconvert}
\alias{bf_convert}
\alias{bfconvert_version}
\title{Convert Bio-Format}
\usage{
bfconvert(
  file,
  outfile = tempfile(fileext = ".tiff"),
  series = NULL,
  timepoint = NULL,
  channel = NULL,
  z = NULL,
  range = NULL,
  tilex = NULL,
  tiley = tilex,
  crop = NULL,
  compression = NULL,
  overwrite = TRUE,
  lookup = TRUE,
  bigtiff = FALSE,
  opts = "",
  verbose = TRUE,
  run = TRUE
)

bf_convert(
  file,
  outfile = tempfile(fileext = ".tiff"),
  series = NULL,
  timepoint = NULL,
  channel = NULL,
  z = NULL,
  range = NULL,
  tilex = NULL,
  tiley = tilex,
  crop = NULL,
  compression = NULL,
  overwrite = TRUE,
  lookup = TRUE,
  bigtiff = FALSE,
  opts = "",
  verbose = TRUE,
  run = TRUE
)

bfconvert_version()
}
\arguments{
\item{file}{File to Convert}

\item{outfile}{Output filename}

\item{series}{All images in the input file are converted by default.
To convert only one series, set this to a numeric}

\item{timepoint}{To convert only one timepoint, set this to a numeric}

\item{channel}{To convert only one channel, set this to a numeric}

\item{z}{To convert only z section, set this to a numeric}

\item{range}{To convert images between certain indices (inclusive),
set 0 to first}

\item{tilex}{width in pixels of each tile}

\item{tiley}{height in pixels of each tile.}

\item{crop}{of the format x,y,width,height. The (x, y) coordinate
(0, 0) is the upper-left corner of the image;
x + width must be less than or equal to the image width
and y + height must be less than or equal to the image height.}

\item{compression}{By default, all images will be written uncompressed.
This is the type of compression done, like LZW.}

\item{overwrite}{Should the file be overwritten}

\item{lookup}{To disable the conversion of lookup tables,
leaving the output file without any lookup tables, set this to
\code{FALSE}}

\item{bigtiff}{This option forces the writing of a BigTiff file}

\item{opts}{Additional options to pass to \code{bfconvert}}

\item{verbose}{Should the command be printed}

\item{run}{Should the command be run?  Useful for diagnostics.}
}
\value{
The output file name
}
\description{
Convert Bio-Format
}
\examples{
file = "~/Downloads/2017_08_03__0006.czi"
if (file.exists(file)) {
res = bfconvert(file = file, run = FALSE)
res
}
}
