% Generated by roxygen2 (4.0.2): do not edit by hand
\name{showinf}
\alias{showinf}
\title{showinf tool extracts metadata for an image}
\usage{
showinf(f, outfile = NULL, ...)
}
\arguments{
\item{f}{Path to input file}

\item{outfile}{Path to output file, or a logical value. When
\code{outfile=TRUE}, a path of \code{/path/to/input.lsm} will result in an
output file of \code{/path/to/input.txt}.}
}
\description{
showinf tool extracts metadata for an image
}

