% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showinf.R
\name{showinf}
\alias{showinf}
\title{Extract metadata for an image using bioformats}
\usage{
showinf(f, ..., outfile = NULL)
}
\arguments{
\item{f}{Path to input file}

\item{...}{Additional character arguments passed to showinf tool.}

\item{outfile}{Character vector specifying path to output file, or a logical 
value. When \code{outfile=TRUE}, a path of \code{/path/to/input.lsm} will 
result in an output file of \code{/path/to/input.txt}. When 
\code{outfile=NULL}, the result will be passed back as a character vector}
}
\value{
A character vector containing metadata or an integer value indicating
  success or failure of of call.
}
\description{
Extract metadata for an image using bioformats
}
