% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showinf.R
\name{showinf}
\alias{showinf}
\alias{showinf.help}
\alias{bf_showinf}
\alias{bf_show_info}
\alias{showinf_version}
\alias{bf_meta_subset}
\title{Displaying images and metadata}
\usage{
showinf(
  file,
  pixel_data = FALSE,
  series = NULL,
  range = NULL,
  crop = NULL,
  autoscale = FALSE,
  ome_xml = FALSE,
  opts = c("-no-upgrade", "-novalid"),
  debug = FALSE,
  verbose = TRUE,
  run = TRUE
)

showinf.help()

bf_showinf(
  file,
  pixel_data = FALSE,
  series = NULL,
  range = NULL,
  crop = NULL,
  autoscale = FALSE,
  ome_xml = FALSE,
  opts = c("-no-upgrade", "-novalid"),
  debug = FALSE,
  verbose = TRUE,
  run = TRUE
)

bf_show_info(
  file,
  pixel_data = FALSE,
  series = NULL,
  range = NULL,
  crop = NULL,
  autoscale = FALSE,
  ome_xml = FALSE,
  opts = c("-no-upgrade", "-novalid"),
  debug = FALSE,
  verbose = TRUE,
  run = TRUE
)

showinf_version()

bf_meta_subset(file)
}
\arguments{
\item{file}{File to display}

\item{pixel_data}{Should the image/pixel data be read in?
If \code{FALSE}, the \code{-nopix} is passed.}

\item{series}{All images in the input file are converted by default.
To convert only one series, set this to a numeric}

\item{range}{To convert images between certain indices (inclusive),
set 0 to first}

\item{crop}{of the format x,y,width,height. The (x, y) coordinate
(0, 0) is the upper-left corner of the image;
x + width must be less than or equal to the image width
and y + height must be less than or equal to the image height.}

\item{autoscale}{Adjusts the display range to the
minimum and maximum pixel values.}

\item{ome_xml}{should this populate OME-XML metadata}

\item{opts}{Additional options to pass to \code{showinf}}

\item{debug}{Enables debugging output if more information is needed}

\item{verbose}{Should the command be printed}

\item{run}{Should the command be run?  Useful for diagnostics.}
}
\value{
The output file name
}
\description{
Displaying images and metadata
}
\examples{
file = "~/Downloads/2017_08_03__0006.czi"
if (file.exists(file)) {
res = showinf(file = file, run = FALSE)
res
}
showinf_version()
}
